[chore][pkg/stanza] Cleanup jsonarray parser operator files
