+++ pkg/stanza/operator/parser/jsonarray/parser_test.go
